backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "radiancefleet"
version = "0.1.0"
description = "Open source maritime anomaly detection for shadow fleet triage"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "shapely>=2.0.0",
    "polars>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "typer>=0.12.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "alembic>=1.14.0",
    "rich>=13.9.0",
    "rapidfuzz>=3.0.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "websockets>=12.0",
    "zstandard>=0.22.0",
    "unidecode>=1.3.0",
    "pyais>=2.5.0",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "PyJWT>=2.10.0",
    "resend>=2.0.0",
    "structlog>=24.0.0",
    "pybreaker>=1.0.0",
    "psycopg2-binary>=2.9",
    "fpdf2>=2.8.0",
    "passlib>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.0",
]
monitoring = ["sentry-sdk[fastapi]>=2.0.0"]
metrics = ["prometheus-fastapi-instrumentator>=7.0.0"]

[project.scripts]
radiancefleet = "app.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501", "E402"]

[tool.ruff.lint.per-file-ignores]
"app/models/*.py" = ["F821"]  # SQLAlchemy Mapped[] forward references

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:xdist -p no:randomly --timeout=30"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest-timeout>=2.4.0",
]
